1ファイル1機能の構成に変更
