+++ app/pages/auth/index.ts
@@ -1,0 +1,5 @@
+import LoginPage from './LoginPage.js'
+
+export {
+  LoginPage,
+}
