+++ app/routes/auth-route.ts
@@ -1,20 +1,20 @@
 import { Hono } from 'hono'
 import {
-  getLoginPage,
-} from '../pages/auth-page.js'
+  LoginPage,
+} from '../pages/auth/index.js'
 import {
-  postPasswordLogin,
-  getGoogleLoginRedirect,
-  getGoogleLoginCallback,
-  getLogout,
-} from '../servers/auth-service.js'
+  loginPassword,
+  loginGoogleRedirect,
+  loginGoogleCallback,
+  logout,
+} from '../servers/auth/index.js'
 
 const authRoutes = new Hono()
 
-authRoutes.get('/', getLoginPage)
-authRoutes.post('/', postPasswordLogin)
-authRoutes.get('/google/redirect', getGoogleLoginRedirect)
-authRoutes.get('/google/callback', getGoogleLoginCallback)
-authRoutes.get('/logout', getLogout)
+authRoutes.get('/', LoginPage)
+authRoutes.post('/', loginPassword)
+authRoutes.get('/google/redirect', loginGoogleRedirect)
+authRoutes.get('/google/callback', loginGoogleCallback)
+authRoutes.get('/logout', logout)
 
 export default authRoutes
